requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
python-dotenv==1.0.1
httpx==0.27.2
google-genai==1.68.0
